refactor: rename views.py to bot.py
